+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Help text clarifications: SAN prefixes, default output directory, bundle behaviour for convert, and TLS retry behaviour
+
 ## [0.4.1] - 2026-09-09
 
 ### Fixed